+++ examples/ta_lib/ta_lib_probe.h
@@ -1,0 +1,1 @@
+#include <ta-lib/ta_libc.h>